+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Enhancements
 
 - Added new parameter `OriginPathPattern` to allow for more flexible origin path matching. `/{stageId}/public` is no longer the required origin path.
+- Added ability to run the Lambda functions in DEBUG without extra AWS SDK logs
 
 ## v0.0.1 - 2026-01-07